+++ src/encoded/types/user_content.py
@@ -1,5 +1,5 @@
 from copy import deepcopy
-from snovault import abstract_collection
+from snovault import abstract_collection, collection
 from encoded_core.types.user_content import UserContent as CoreUserContent
 from .base import Item as SMAHTItem
 from .base import mixin_smaht_permission_types
@@ -11,8 +11,8 @@ ENCODED_CORE_USER_CONTENT_SCHEMA = deepcopy(CoreUserContent.schema)
 @abstract_collection(
     name='user-content',
     properties={
-        'title': 'Static Sections',
-        'description': 'Static Sections for the Portal',
+        'title': 'User Content',
+        'description': 'User Content for the Portal',
     })
 class UserContent(SMAHTItem, CoreUserContent):
     item_type = 'user_content'